pyproject.toml
[project]
name = "grepmail"
version = "0.1.0"
description = ""
authors = [
    {name = "Kanak Tanwar",email = "kanaktanwarpro@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mindsdb-sdk (>=3.4.3,<4.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[tool.poetry.scripts]
grepmail = "grepmail.main:app"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
